Move 'Drill scaling up number of workers' above 'grouped' drills
This was an outlier, as the other drills below it are all
composite drills made up of at least two drill tasks. I think it's
easier to maintain by keeping all the self-contained drills at the
top.

NB: I suspect this drill can be deleted post-replatforming, but it
is still worth keeping for now.