Revise policy on semver RE dropping ruby versions
This softens our previously documented stance: we now encourage releasing the dropping of EOL Ruby versions as minor gem upgrades, as opposed to major upgrades. This follows a [discussion on Slack](https://gds.slack.com/archives/CAB4Q3QBW/p1680164984754749) where the consensus was that it should be changed, for the reasons outlined in this document.

Note that I've also standardised the markdown links approach to use aliases as per the rest of this section, and have reordered the aliases alphabetically.