Remove 'get access to GitHub' instructions from Get Started
Now link to the canonical page instead, so we can keep the policy
documented in one place.

This exposed a slightly awkward instruction later on in the docs,
whereby one's `ssh_username` should be added to the user reviewer
(in a second PR, if you're following the instructions to the letter).
In practice, devs tend to take inspiration from previous PRs, most
of which have both changes being made at the same time, so I've
added a warning/suggestion here.