Frontload the "What you need to do" instructions for new repos
The previous version listed all the repo requirements, THEN a list
of what you need to configure on your new repo, but it was also
muddled in with a section about things govuk-saas-config takes
care of for you. The risk is that devs will misread this and
wrongly infer (as I did!) that things like write access to the
GOV.UK team are handled automatically by govuk-saas-config.

I've now moved the "what you need to do" bit to the top, so that
there is less chance for human error. I've also fixed an inconsistency
whereby write access for the GOV.UK team was omitted from the
'rules' section, presumably by mistake.