Update architecture diagram
The architecture diagram was looking a little out of date.

In particular:
- Mapit instead of locations-api
- manuals-frontend still there even though it's been retired
- info-frontend making calling out to performance platform (which has been retired)

I fixed these things in the source diagram, but that seems to have
broken the original embedding code (draw.io has moved to diagrams.net,
so I'm not surprised the embed broke).

This commit updates the embed, and attempts to size the diagram so it
takes up more of the screen.

My CSS is a bit hacky, but the idea is "the whole view width, minus the
sidebar and margins, appart from on mobile where it's the whole view
width minus the margins". The 700px height is arbitrary. Seems to just
about work.