Add clarification about MFA device name
Two new team members have separately made the same assumption, and came
up against the same error, when naming our MFA devices in AWS.

The assumption was that "must be prefixed" meant the name could, and
maybe even should, also have a suffix.

However, we couldn't authenticate in the AWS console on the command line
with the codes generated by devices with suffixed names. The only way to
successfully authenticate was to add an MFA device and name it with the
exact string of the email address.