Add govuk_frontend_toolkit_gem repo
This repo should be treated the same (in terms of ownership) as
govuk_publishing_components (as that is the successor), which
would imply an owner of "#govuk-frontenders". However, being a
deprecated repo, from a quick look, it is only used in Whitehall.
This therefore means the only team with agency to archive the repo
is the owning team for Whitehall:
"#govuk-publishing-experience-tech".

See conversations at https://github.com/alphagov/govuk-developer-docs/pull/3913/files#r1132283314
followed by https://gds.slack.com/archives/CARGH33JS/p1686667844288349

Trello: https://trello.com/c/tsBL9lPH/3161-work-through-list-of-missing-repos