Add pay-pagerduty repo
This has been used for the last two quarters, to synchronise our
rotas with PagerDuty.
It has Pay in the name, but GOV.UK are slowly in the process of
adopting it.

Its owner is #govuk-platform-security-reliability, following the
precedent of other '2nd line leads' type repos such as
govuk-rota-generator.

Trello: https://trello.com/c/tsBL9lPH/3161-work-through-list-of-missing-repos