Support multiple apps per repo, via new `app_name` property
We have a [script](https://github.com/alphagov/govuk-saas-config/pull/142)
that compares the list of repos in the Developer Docs with all of
the alphagov GitHub repos tagged with the 'govuk' topic. It is
currently highlighting this:

```
Untagged govuk repos:
licensify-admin
licensify-feed
```

This is due to a limitation in our current data modelling, which
equates `repo_name` with the application name, and makes a false
assumption that there can only be 0-1 apps per repository.
The [licensify](https://github.com/alphagov/licensify) repo has
3 apps.

Trello: https://trello.com/c/tsBL9lPH/3161-work-through-list-of-missing-repos