Remove retired repos from 'deployable applications' check
These repos were specified in a hardcoded `intentionally_missing`
array, to avoid the `verify_deployable_apps` task from shouting
about apps in Puppet that were missing from the Developer Docs.

This commit removes those retired repos from the array and adds
or updates their entry in the repos.yml file to provide a bit of
detail about what they used to do.

NB, we'll probably want to remove or rewrite the
`verify_deployable_apps` task now that apps are on EKS, but that
is out of scope for this PR.