Add .govuk_automerge_config.yml
This is the filename that the auto-merge service will look for
when deciding whether or not a Dependabot PR can be auto-merged.

This is the first draft of the YAML structure too.

The `auto_merge` array is flexible enough to provide granular
control on a per-dependency basis, as each dependency is
represented as a hash, e.g. some dependencies may be safe to
auto-merge patch releases only, whereas others may want to
allow minor releases. The hash structure also leaves room for
us to add additional constraints in future if we wish.

The `api_version` property is intended to be a safeguard that
allows us to tailor the automerge service based on when a
particular repo opted in to the automerge behaviour. For instance,
we could look at the `api_version` and offer backwards compatibility
as we change the API, or we could fail gracefully and give the
user some information as to how to upgrade to a supported version.

I don't really intend for this to ever have to be updated, but
specifying an `api_version` will give the future maintainers of
the automerge service a way of introducing breaking changes
in a controlled way.

While the service is being developed, I'll keep the version as
`0`. Once the initial service is feature-complete as per RFC-156,
I'll set the version to `1`.

Trello: https://trello.com/c/RblylctX/3136-build-the-auto-merge-service-for-dependabot-prs-5