Add tech lead as qualifier to Get Started
Since access to a tech lead is required in a few steps I felt it'd be
better to deal with the need for one as a pre-requisite where we could
also outline the senior tech escalation route without needing to
reference it multiple times.

I've updated the other references to asking for help to be ask tech lead
as I think that is simpler for someone new to have one clear point of
contact and less decisions to make.

I also fixed some of the inconsistent line lengths in the signon
sections.