Add a Slack profile step in Get Started
If we require people to set-up their Slack profile we should have that
as a distinct step rather than a pre-requisite.

I've also clarified that we expect people to fill in the title field, as
Slack provides a "role" field that is only visible when accessing
external Slack instances.

I've also removed the "GitHub handle" and "why you need access" as both
of these are things not routinely added to a Slack profile and thus a
source of confusion.

My motivation for this change is working with a new starter who missed
the initial instruction and then filled in the role field once they
followed them.