Resolve unnecessary absolute links
This is a find and replace which updates internal links in the developer
docs to be host and scheme independent.

The motivation for this was that I got frustrated at clicking links when
developing locally and ending up out of my local version of dev docs and
into the production one.