Move AWS IAM step before "get permissions for AWS, GitHub, etc"
Prior to this commit, users may have followed the GitHub steps
(previously step 4, now step 5) and then jumped ahead to try to
get added to the alphagov org. We had a special instruction in
step 6 to make sure you create your AWS IAM account first, so
presumably that skipping has happened in the past!

Now reordered the steps such that the AWS IAM account is created
first, then the general GitHub steps, followed by getting access
to the GitHub org. It also reinforces that steps 1-5 are required
before you can be added to the org.