Include content schemas on document type page
This adds a new section to a document type page that specifies which
schemas support the document type. The motivation for adding this
information was a request from Sally in the Content Types Advisory Group
for ease of establishing which schema a document_type corresponds to.

The implementation for this relies on a couple of assumptions that could
eventually fall. One is that each schema has at least a notification
schema (which is currently true) - this is required because the `.find`
method on GovukSchemas::Schema requires a schema type to be given and
will raise an exception if it can't find a schema. So it was a choice
between a) editing the gem b) having a delicate set of begin ... rescue
or c) go with the assumptive approach. The other assumption is that
document types are always defined in schemas in an enum structure. This
seems pretty safe because this isn't user configurable in Publishing API
content schemas [1].

The end result of this work was a little underwhelming, there are a
number of schemas that allow any document type which means the list can
seem bizarre and overwhelming. I've tried to resolve most of these with
a Publishing API PR [2] which resolves most of these except for
"generic", "generic_with_external_links" and "placeholder".

[1]: https://github.com/alphagov/publishing-api/blob/b0cf0ebd9c25b1bb918fb9491c93a3118fb2599e/lib/schema_generator/format.rb#L187-L189
[2]: https://github.com/alphagov/publishing-api/pull/2466