Retire GitHub pages sites carefully
GitHub pages site commonly use URLs like this: https://alphagov.github.io/.
If we retire a standard site like this (that has no custom DNS configuration
on our side), then we should be mindful to update our docs and code beforehand
because it would be a nuisance if something broke because it relied on this.

We should also be a little careful because we don't own the github.io domain. It
could change ownership at some point, and we don't want to have to go
through our archived repos to clean everything up when we could have done
it at the time.

If we use a custom domain and disable the site, we're at risk of a domain
takeover. That is: someone else could host a site using our URL.

We can protect against this by using [domain verification](https://docs.github.com/en/pages/configuring-a-custom-domain-for-your-github-pages-site/verifying-your-custom-domain-for-github-pages)
and/or checking for dangling CNAMEs, but it's better if we're not totally
reliant on these features.

It's also great to keep our DNS config tidy!