Shift low value schemas to bottom of list
The generic and generic_with_external_links schemas are likely
safe to be deleted. As per Kevin's analysis, there is zero evidence
of usage, however there is code in Mainstream Publisher (which we
don't think is executed), which muddies the water slightly:
https://github.com/search?q=repo%3Aalphagov%2Fpublisher%20GenericEditionPresenter&type=code

The 'placeholder' schema is also of limited value. We've therefore
agreed to shift these low value schemas to the bottom of the list
of content schemas.