Document using a read-only sandbox
This isn't the best long-term place for this documentation, but
for the time being, if this is where developers will be looking
to figure out how to get a console on their apps, it makes sense
to nudge them towards using safer approaches.

They may also be tripped up by the double `-- --` in the command -
one to pass an argument to the rails console, and the other to
pass an argument to kubernetes. With that in mind, this feels like
a valuable thing to document.

Note that whilst a read-only console will roll back any changes
made, it won't roll back _everything_, e.g. if you call code that
enqueues jobs or sends email. So devs still need to be sensible.