Create "multipage_layout" for multi_page_table_of_contents behaviour
This commit contains the minimum set of changes to retain the
hierarchical/nested structure of the Kubernetes docs (which
originally made use of the ["multipage documentation" feature of
the Tech Docs Template](https://tdt-documentation.london.cloudapps.digital/configure_project/structure_docs/#build-a-multipage-documentation-site)),
without breaking all of our other docs. This is achieved by adding
a new layout that closely mimics the multipage documentation feature,
which we will opt the '/kubernetes' folder into in the next commit.

We're unable to use the native feature provided by the template,
for several reasons:

1. The existing documentation is massive in scale, and flat in
   structure. To avoid making the sidebar navigation too messy,
   a significant piece of work would need to be done to rearchitect
   the current docs, choose appropriate page weights, group
   related docs into nested folders, and so on.
2. A lot of documentation is pulled in externally at compile time
   - a great proportion of documentation doesn't actually live in
   the Developer Docs repo. The mutlipage doc feature has a
   requirement that every single page has an explicit h1 tag, which
   we could add relatively easily to the Developer Docs' "local"
   documentation, but would be a significant undertaking to roll
   out across all of the external repositories whose 'docs'
   folders we import. NB, currently we derive the h1 from each
   document's 'title' frontmatter, which has been working just
   fine and avoids duplication.
3. There is an [/analytics microsite](https://docs.publishing.service.gov.uk/analytics/)
   hosted on the Developer Docs, which has its own structure and
   layout, and which does not play nicely when enabling the
   multipage feature.

Note that we've had to manually invoke the JavaScript that styles
the link representing the 'current page' in the sidebar. Presumably
this doesn't happen automatically because we haven't switched the
whole site to multipage mode, for the reasons outlined above.
The JS has been written quite defensively so that if it fails, it
won't interfere with any other JS, and the worst case scenario is
that we lose some styling.