Document setting up a new GOV.UK repo
We had a page for this (source/manual/configure-github-repo.html.md) but it was removed in #4137 in an effort to consolidate our GitHub documentation. This commit better structures the revised docs, to make it clear this is a list of steps to follow, and to make it easier to find in search.