Update to refer to govuk-fastly
How we configure the `/alerts` endpoint has changed in the move to
govuk-fastly - see changes for details.

There's also no longer any 'merging' of Fastly UI config with our
VCL, which we previously had in place to enable POP shielding on
the /alerts endpoint. That is now natively supported in our move
to Terraform.