Add a step to 'Transition a site to GOV.UK' docs
The bouncer deploy script was migrated to terraform and now doesn't run
automatically when a subdomain is added to transition-config. There is also an
additional domain verification step.

Given the plans to consolidate transition-config into the transition app,
updating a doc seems sufficient at this stage.  The product team can later
decide on automation of this process.