Remove AWS Backup section from backups doc
Removing this section as it was only aspirational and does not reflect any of the backups we actually have implemented.

Platform Security & Reliability [investigated using AWS Backup](https://trello.com/c/EENvDlD4/3229-decide-how-we-should-use-aws-backup-vaults) but concluded that the engineering effort would be quite considerable, and that we could get many of the benefits of AWS Backup by making relatively modest changes to our existing setup. We may revisit this in future, but for now, removing the section.