Move kubernetes-infrastructure.html -> /kubernetes/cheatsheet.html
The original file was created prior to #4186, which imported all
of the Kubernetes docs from the old
govuk-kubernetes-cluster-user-docs site.

It now makes sense to either delete the quick reference guide, or
consolidate it into the '/kubernetes' folder.
Some engineers may still be relying on the quick reference guide,
so I've opted for the latter.