Update the checklist for archiving a GitHub repo.
Try to reduce the burden involved in archiving a repo (since we want
people to archive disused stuff and not be put off by a big toily
process) by clarifying that there are situations where e.g. it may be
overkill to add a signpost to the README (in addition to the banner that
GitHub adds anyway).

Also simplify some of the wording and make it a bit clearer as to why
the remaining toil is needed.