Remove confusion between Logit and Kibana
We currently have documentation for using Logit split across two pages
with some duplicated information. Therefore merging both so there is a
single guide without duplication.

Also updating references to 'Kibana' throughout the docs to make it
clear this is an interface for querying the logs we have shipped to
Logit.

Using two names for the same tool has made it difficult to search for
documentation, so ensuring the words "Logit" and "Kibana" appear
alongside each other whenever referenced.