Update ZenDesk leaver ticket process
Add a new comment advising that the ticket can be closed if no reference is found when searching the `govuk-user-reviewer` repository.

This should help speed up the process when dealing with leavers tickets as the only other action to do in this scenario is check "https://github.com/orgs/alphagov/teams/gov-uk-content-designers/members", for which we already have automated checks in place for.