Add information about including code snippets as Zendesk notes
This updates the Zendesk guidance to state that you should include a
copy of any commands that are executed directly on a console as
internal notes on Zendesk tickets.

By including the commands, it helps the next shift (and/or product
teams) understand what has already been attempted and identify any
possible problems that could arise.