Update getting started
As a new developer onboarding in February 2024 I got caught out on a
number of items, this one caused me the most pain.

You have to set the MFA device name in AWS to your email address for the
GDS-CLI to work later - it is possible to workaround this, but as a new
dev joining GDS, it is possible best to make this appear necessary!

Reading the docs again, this requirement is mentioned, but I certainly
missed it!

This commit adds a couple of tweaks that might help others.