Update Kubernetes getting started
As a new developer onboarding in February 2024 this one caused me some
pain.

Not setting the default namespace meant none of the example
`kubectl` commands worked as they are all missing the required `-n apps`
adding it early in the getting started section seems like a solid plan
to help those that will follow.