Remove unneeded `dashboard_url: false` properties
These were only set because the default behaviour used to be to
return a guessed dashboard URL, which would have been incorrect
for these repos. So `false` was set to explicitly prevent that
behaviour.

The behaviour now defaults to not returning anything, unless the
app is hosted on EKS. So these properties are no longer required.