Update "Request an AWS account" URL
The GDS Request an AWS Account Tool is being migrated off PaaS this month.

From a user perspective, the tool will still function exactly as before, although the URL of the tool will change. For the time being, the current URL for the tool will redirect to the new URL, which is:
https://request-an-aws-account.gds-reliability.engineering

The change is going live by the 26th February.

Via email from Patrick Berkeley