Remove link to database engines
This link was to a point in time document from 2022 and is not maintained. The information in it is outdated and risks being misleading as time continues to pass by. I'm not aware of an alternative document that is kept up-to-date (the GOV.UK dependencies sheet doesn't have the DB engine specificity https://docs.google.com/spreadsheets/d/137KZhjctJ8qTKYPnq2QNVkyoIC6ok7KA2G1vErNC6Oo/edit#gid=0)