Update deploy code freeze documentation
While drilling deploying a code freeze on 2ndline we noticed that
this set of documentation is now out of date.

While you previously just had to check the Freee deployments checkbox
on the applications edit page on release, you now need to manually run
a GitHub Action to toggle off CD.

This updates the docs to reflect this.