Update dependencies + work around govuk_tech_docs.
Update all the Rubygems dependencies as far as possible.

Had to specify a couple of indirect dependencies missing from
govuk_tech_docs and avoid html-pipeline >= 3, which appears to be
incompatible with govuk_tech_docs for now.

Group the dev/test-only dependencies in the Gemfile so we can skip them
in the build later.