Fix site search
Site search in the developer docs was defaulting to use Google search following a recent update in `modules.js` in the publishing-components-gem.

It appears this was conflicting with the approach used in the developer docs where jQuery is still used, and modules.js would initialise all modules on the page, even where jQuery is used.

This change copies over `modules.js` and `dependencies.js` from the publishing components gem.

`modules.js` has been modified to add support for initialising modules that rely on jQuery