Update policy on staffing 2nd line during away days
2nd line is no longer the full time job it used to be: the platform is much more stable, and most application support work is now being handled directly by the product teams. 2nd line is therefore largely an incident response team now.

It therefore seems difficult to justify someone's exclusion from away days, if most of the time there is very little for 2nd line to do. So the updated guidance allows it, on the proviso that the 2nd line engineers regularly check Zendesk and continue to be interruptible by PagerDuty. In the first instance, we would also ask that engineers try to find cover.