Move Elasticsearch alerts docs under Infrastructure
There's no longer an Incinga instance checking the health of the
elasticsearch cluster for us, but a lot of this documentation is still
salvageable.

I've been lazy and not fixed the guidance on tunneling, instead just
linking to a similar idea for OpenSearch. We should probably make that
documentation more generic, as it's already used for Manage Amazon MQ
and Manage OpenSearch on AWS.