Remove redundant docs on email alert unprocessed work alert
These alerts no longer exist, and most of this documentation is talking
about what the alerts are, and what they mean.

There are some specific suggestions at the end of the doc, but they
involve running things in the rails console which is a very icky anyway.
The rest of the documentation is generic enough I'm confortable deleting
it.