Ignore content_block_email_address schema
This schema uses a `type` of `email_address`, something which is not
supported in older versions of `govuk_schemas`. We can’t bring in
the new version yet, because of dependency issues with Middleman.

We have an [upstream PR](https://github.com/middleman/middleman/pull/2709)
to fix this, but in the interests of unblocking things, we can safely
ignore this new schema for now.