Link to the Argo CD logs section
This sentence instructs you to check the Argo logs but if you're not
familiar with Argo, you'll want to know how to do that.

So this adds a link to the section of this documentation page describing
how to check the logs. It's immediately below the section I've added the
link to but if you're just scanning the page you might not realise that.
Also, if some additional documentation is inserted between these
sections in the future, it'll be useful to be able to jump straight to
the logs section.