Add support for custom 'kibana_url' / 'kibana_worker_url' fields
These fields are already referenced in the tests (though not
currently used anywhere in repos.yml) as a way of 'unsetting' the
URLs. But the convention elsewhere in govuk-developer-docs is that
such fields provide a way of specifying the value (which may be
`false`, to unset the value, or may be a valid string).