Clarify the wording in 'When you get Production Deploy access'
The term “Senior Technologist” can be ambiguous as it can be interpreted as
referring to any Senior Developers/SRE/Technical Architects. Also later in the
doc “Apprentice Technologist” is used.

We commonly use phrase “Senior Tech” to describe the group of Lead
Developers/SRE/Technical Architects. This also adds a link to the doc
describing how to contact that group.