Document rolling back by version tag rather than SHA hash
From chatting with Helen about the roll back process we couldn't understand why we'd need the SHA hash when we could use the simpler version id.

I've optimistically assumed that this information is outdated and we can use the version tag and thus updated the documentation. Will seek a review from someone more in the know.