Document how to use 2fa-seed strings
It took some digging to find this from the old govuk-secrets
instructions. We don't need it often, but it's useful to know how
to do this for the odd occasion we need to sign into a shared
account.

I've documented a lean, one-time-use way of doing it, and also
a more involved but persistent way of doing it.