Rename `dependencies_team` to `alerts_team`
This value was used by the (now deprecated) `govuk-dependencies`
application to allow dependabot alerts to be posted to channels other
than the team's main channel.

It is currently unused by any application, but we would like to alert to
a different channel in the release app.

Therefore renaming the field to make it more generic, so it can be used
by all alerts, not just dependencies.