Remove alert doc about Pingdom homepage check
We've had a number of PagerDuty alerts about failing Pingdom
checks recently, e.g.
https://governmentdigitalservice.pagerduty.com/incidents/Q2P01JX296D7O7

The alert hasn't linked to this guidance, and it doesn't seem as
though developers have gone looking for this guidance either. The
scenarios outlined in the doc are pretty unlikely, and would be
fairly straightforward to come to a conclusion on without the doc,
so there's no value in keeping this.