Document updating search autocomplete denylist
This provides information on how to perform the relatively manual task
of updating the list of terms that are forbidden from showing for the
GOV.UK search autocomplete.

The motivation for adding this to the developer docs is that this is a
part of GOV.UK that is due to go live in the coming days and if the
autocomplete starts showing something embarrassing it could be an urgent
task to update it - which may need to happen out-of-hours. Due to this
it seemed best to be easily searchable from the developer docs.

It is expected that this manual process will be replaced by software
that can perform this task in Q4 of 2024-2025 - so hopefully this
documentation won't be needed for long.

Note: I wanted to first put this manual in a section called "Search" but
found this had a conflict as the anchor "#search" was already used by
the search box. I went for GOV.UK Search as a way to create a unique id.