Documentation on disabling search autocomplete
This explains the steps that are required to disable autocomplete for
the GOV.UK Site Search. It is expected that once we've had this feature
on GOV.UK for a few weeks / months we will have sufficient confidence in
it that we will no longer need to disable it and thus can remove this
functionality.

The most likely reason that we think this feature might be needed is if
there is an urgent denylist problem and an out of hours engineer
struggles with the manual process for updating it.