Add section on Whitehall bugs
This is a temporary section. We are addressing the root cause of the bugs.
A data patch has also been run. Nonetheless, outliers might exist, that we
will tackle on 3rd line. This guidance can thus come in handy.

I thought about putting this in whitehall, but I feel like it best belongs here,
albeit at the bottom of the file.