♻️ Include overview diagram and additional sections
This commit hopes to improve the understanding of how our internal platform works. By breaking things down into component level detail we should be able to tell a story of the current platform model and help gain a shared understanding of what's currently running. This commit doesn't finish the story and there is plenty more to add to this document.