Add combined repos by team and type breakdown
It's easy to find all of a team's apps or all of a team's repos using
the by team sections of the apps and repos pages. In Publishing Platform
we recently had a need to find all of our gems (perhaps when looking to
remove support for Ruby 3.1, which is reaching end of life). This was
less easy. We could compare our repos by team with the list of gems in
repos by type, or try to use our memory with either list, but neither
solution is ideal, and the latter wouldn't work very well for people
outside of or new to our team

This adds a combined breakdown of repos by team and type to help in such
situations. Given the use cases for this are probably quite niche, I've
added it as a separate page linked from the main repos page, rather than
clogging up an already-complex page

I'd like to be able to add breadcrumbs for easy navigation back to the
main repos page, but it's proving challenging to get this working
properly in a html.md.erb file. Adding
`GovukPublishingComponents.render` initially blew away all subsequent
Markdown content. Adding `.strip` to the end per
`source/templates/repo_template.html.md.erb` fixed that, but the spacing
was off: there was a gap between the breadcrumbs and the rest of the
content such that the rest of the content was lower than the table of
contents. It's not worth the effort to dedicate too much time to right
now, so I've just added a link back to the repos page in an opening
paragraph