WIP add "services" which repos depend on
The idea would be for there to be some easier way for us to say "which
apps do we have that use postgres", or "which apps do we have that use
memcached". It would be nice if we could do this automatically, but I'm
not sure if there's a great datasource for it - kubernetes maybe?